// generated/models/ApiCollectTdhTargetItem.ts
// @ts-nocheck
/**
 * 6529.io API
 * JSON REST API for 6529.io. New here? A terminology overview and a step-by-step authentication walkthrough (guides &amp; auth) live at <a href=\"https://6529.io/tools/api\">https://6529.io/tools/api</a>. The raw machine-readable spec is downloadable at <a href=\"/openapi.yaml\">/openapi.yaml</a> and <a href=\"/openapi.json\">/openapi.json</a>.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { ApiCollectAsset } from '../models/ApiCollectAsset';
import { ApiMarketTradeOrder } from '../models/ApiMarketTradeOrder';
import { HttpFile } from '../http/http';

export class ApiCollectTdhTargetItem {
    'asset': ApiCollectAsset;
    'order': ApiMarketTradeOrder;
    'quantity': string;
    'recipient': string;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "asset",
            "baseName": "asset",
            "type": "ApiCollectAsset",
            "format": ""
        },
        {
            "name": "order",
            "baseName": "order",
            "type": "ApiMarketTradeOrder",
            "format": ""
        },
        {
            "name": "quantity",
            "baseName": "quantity",
            "type": "string",
            "format": ""
        },
        {
            "name": "recipient",
            "baseName": "recipient",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return ApiCollectTdhTargetItem.attributeTypeMap;
    }

    public constructor() {
    }
}